Add unit tests for ledger backup and minimum visits
